+++ tests/unit/lib/NavLink-test.js
@@ -20,6 +20,7 @@ var navigateAction = require('../../../').navigateAction;
 var TestRouteStore = RouteStore.withStaticRoutes({
     foo: { path: '/foo', method: 'get' },
     bar: { path: '/bar', method: 'get'},
+    int: { path: '/internal', method: 'get'},
     fooA: { path: '/foo/:a', method: 'get' },
     fooAB: { path: '/foo/:a/:b', method: 'get' }
 });
@@ -31,24 +32,36 @@ onClickMock = function () {
 describe('NavLink', function () {
     var mockContext;
 
-    beforeEach(function () {
-        global.document = jsdom.jsdom('<html><body></body></html>');
-        global.window = global.document.parentWindow;
-        global.navigator = global.window.navigator;
-        React = require('react');
-        ReactDOM = require('react-dom');
-        ReactTestUtils = require('react-addons-test-utils');
-        mockContext = createMockComponentContext({
-            stores: [TestRouteStore]
+    beforeEach(function (done) {
+        jsdom.env({
+            url: 'http://yahoo.com',
+            html: '<html><body></body></html>',
+            done: function (err, window) {
+                if (err) {
+                    done(err);
+                    return;
+                }
+                global.document = window.document;
+                global.window = window;
+                global.navigator = window.navigator;
+
+                React = require('react');
+                ReactDOM = require('react-dom');
+                ReactTestUtils = require('react-addons-test-utils');
+                mockContext = createMockComponentContext({
+                    stores: [TestRouteStore]
+                });
+                mockContext.getStore('RouteStore')._handleNavigateStart({
+                    url: '/foo',
+                    method: 'GET'
+                });
+                MockAppComponent = require('../../mocks/MockAppComponent');
+                NavLink = require('../../../lib/NavLink');
+                createNavLinkComponent = require('../../../lib/createNavLinkComponent');
+                testResult = {};
+                done();
+            }
         });
-        mockContext.getStore('RouteStore')._handleNavigateStart({
-            url: '/foo',
-            method: 'GET'
-        });
-        MockAppComponent = require('../../mocks/MockAppComponent');
-        NavLink = require('../../../lib/NavLink');
-        createNavLinkComponent = require('../../../lib/createNavLinkComponent');
-        testResult = {};
     });
 
     afterEach(function () {
@@ -69,6 +82,7 @@ describe('NavLink', function () {
             expect(ReactDOM.findDOMNode(link).getAttribute('href')).to.equal('/foo');
             expect(ReactDOM.findDOMNode(link).textContent).to.equal('bar');
         });
+
         it('should prefer href over routeName', function () {
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
@@ -77,6 +91,7 @@ describe('NavLink', function () {
             );
             expect(ReactDOM.findDOMNode(link).getAttribute('href')).to.equal('/foo');
         });
+
         it('should create href from routeName and parameters', function () {
             var navParams = {a: 1, b: 2};
             var link = ReactTestUtils.renderIntoDocument(
@@ -86,6 +101,16 @@ describe('NavLink', function () {
             );
             expect(ReactDOM.findDOMNode(link).getAttribute('href')).to.equal('/foo/1/2');
         });
+
+        it('should have routeName mapped to the correct path', function () {
+            var link = ReactTestUtils.renderIntoDocument(
+                <MockAppComponent context={mockContext}>
+                    <NavLink routeName='int' />
+                </MockAppComponent>
+            );
+            expect(ReactDOM.findDOMNode(link).getAttribute('href')).to.equal('/internal');
+        });
+
         it('should throw if href and routeName undefined', function () {
             var navParams = {a: 1, b: 2};
             expect(function () {
@@ -96,15 +121,36 @@ describe('NavLink', function () {
                 );
             }).to['throw']();
         });
+        it('should create href with query params', function () {
+            var queryParams = {a: 1, b: 2};
+            var link = ReactTestUtils.renderIntoDocument(
+                <MockAppComponent context={mockContext}>
+                    <NavLink routeName='foo' queryParams={queryParams} />
+                </MockAppComponent>
+            );
+            expect(ReactDOM.findDOMNode(link).getAttribute('href')).to.equal('/foo?a=1&b=2');
+        });
+        it('should set style and className properties', function () {
+            var link = ReactTestUtils.renderIntoDocument(
+                <MockAppComponent context={mockContext}>
+                    <NavLink routeName='foo' className='foo' style={{
+                        backgroundColor: '#000000'
+                    }} />
+                </MockAppComponent>
+            );
+            expect(ReactDOM.findDOMNode(link).getAttribute('class')).to.equal('foo');
+            expect(ReactDOM.findDOMNode(link).getAttribute('style')).to.contain('background-color');
+        });
         it('should set active state if href matches current route', function () {
             var navParams = {a: 1, b: 2};
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
-                    <NavLink routeName='foo' />
+                    <NavLink routeName='foo' activeClass='active' />
                 </MockAppComponent>
             );
             expect(ReactDOM.findDOMNode(link).getAttribute('class')).to.equal('active');
         });
+
         it('should set active state by tag name if the optional activeElement property is set', function () {
             var navParams = {a: 1, b: 2};
             var link = ReactTestUtils.renderIntoDocument(
@@ -114,6 +160,7 @@ describe('NavLink', function () {
             );
             expect(ReactDOM.findDOMNode(link).nodeName.toLowerCase()).to.equal('span');
         });
+
         it('should set active state with custom class and style', function () {
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
@@ -121,8 +168,9 @@ describe('NavLink', function () {
                 </MockAppComponent>
             );
             expect(ReactDOM.findDOMNode(link).getAttribute('class')).to.equal('bar');
-            expect(ReactDOM.findDOMNode(link).getAttribute('style')).to.equal('color:red;');
+            expect(ReactDOM.findDOMNode(link).getAttribute('style').replace(/ /g, '')).to.equal('color:red;');
         });
+
         it('should set the active state and keep the passed props', function () {
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
@@ -131,21 +179,37 @@ describe('NavLink', function () {
                 </MockAppComponent>
             );
             expect(ReactDOM.findDOMNode(link).getAttribute('class')).to.equal('bar active2');
-            expect(ReactDOM.findDOMNode(link).getAttribute('style')).to.equal('background:blue;color:red;');
+            expect(ReactDOM.findDOMNode(link).getAttribute('style').replace(/ /g, '')).to.equal('background:blue;color:red;');
         });
+
         it('should not set active state if href does not match current route', function () {
             var navParams = {a: 1, b: 2};
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
-                    <NavLink routeName='fooAB' navParams={navParams} />
+                    <NavLink routeName='fooAB' navParams={navParams} activeClass='active' />
                 </MockAppComponent>
             );
             expect(ReactDOM.findDOMNode(link).getAttribute('class')).to.equal(null);
         });
+        
+        it('should able to get additional child props by dynamical getDefaultChildProps function', function () {
+            var navLink = React.createElement(createNavLinkComponent({
+                getDefaultChildProps: function () {
+                    return {'data-foo': 'foo', 'data-bar': 'bar'};
+                }
+            }), {routeName: 'foo'});
+            var link = ReactTestUtils.renderIntoDocument(
+                <MockAppComponent context={mockContext}>
+                    {navLink}
+                </MockAppComponent>
+            );
+            expect(ReactDOM.findDOMNode(link).getAttribute('data-foo')).to.equal('foo');
+            expect(ReactDOM.findDOMNode(link).getAttribute('data-bar')).to.equal('bar');
+        });
     });
 
     describe('dispatchNavAction()', function () {
-        it ('use react context', function (done) {
+        it('use react context', function (done) {
             var navParams = {a: 1, b: true};
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
@@ -162,7 +226,7 @@ describe('NavLink', function () {
                 done();
             }, 10);
         });
-        it ('should getNavParams from overwriteSpec if so configured', function (done) {
+        it('should getNavParams from overwriteSpec if so configured', function (done) {
             var navParams = {a: 1, b: true};
             var params = {
                 href:'/foo',
@@ -189,7 +253,7 @@ describe('NavLink', function () {
                 done();
             }, 10);
         });
-        it ('stopPropagation stops event propagation', function (done) {
+        it('stopPropagation stops event propagation', function (done) {
             var propagateFail = function(e) {
                 expect(e.isPropagationStopped()).to.eql(true);
             };
@@ -206,6 +270,7 @@ describe('NavLink', function () {
                 done();
             }, 10);
         });
+
         it('context.executeAction called for relative urls', function (done) {
             var navParams = {a: 1, b: true};
             var link = ReactTestUtils.renderIntoDocument(
@@ -223,7 +288,68 @@ describe('NavLink', function () {
                 done();
             }, 10);
         });
-        it ('context.executeAction called for routeNames', function (done) {
+
+        it('context.executeAction called for href when validate is true and href is registered', function (done) {
+            var navParams = {a: 1, b: true};
+            var link = ReactTestUtils.renderIntoDocument(
+                <MockAppComponent context={mockContext}>
+                    <NavLink validate={true} href='/internal' navParams={navParams}/>
+                </MockAppComponent>
+            );
+            ReactTestUtils.Simulate.click(ReactDOM.findDOMNode(link), {button: 0});
+            window.setTimeout(function () {
+                expect(mockContext.executeActionCalls.length).to.equal(1);
+                expect(mockContext.executeActionCalls[0].action).to.equal(navigateAction);
+                expect(mockContext.executeActionCalls[0].payload.type).to.equal('click');
+                expect(mockContext.executeActionCalls[0].payload.url).to.equal('/internal');
+                expect(mockContext.executeActionCalls[0].payload.params).to.eql({a: 1, b: true});
+                done();
+            }, 10);
+        });
+
+        it('context.executeAction not called for external href when validate is true', function (done) {
+            var link = ReactTestUtils.renderIntoDocument(
+                <MockAppComponent context={mockContext}>
+                    <NavLink href='/external' validate={true}/>
+                </MockAppComponent>
+            );
+            ReactTestUtils.Simulate.click(ReactDOM.findDOMNode(link), {button: 0});
+            window.setTimeout(function () {
+                expect(testResult.dispatch).to.equal(undefined);
+                done();
+            }, 10);
+        });
+
+        it('context.executeAction called for external href when validate is false', function (done) {
+            var link = ReactTestUtils.renderIntoDocument(
+                <MockAppComponent context={mockContext}>
+                    <NavLink href='/external'/>
+                </MockAppComponent>
+            );
+            ReactTestUtils.Simulate.click(ReactDOM.findDOMNode(link), {button: 0});
+            window.setTimeout(function () {
+                expect(mockContext.executeActionCalls.length).to.equal(1);
+                expect(mockContext.executeActionCalls[0].action).to.equal(navigateAction);
+                expect(mockContext.executeActionCalls[0].payload.type).to.equal('click');
+                expect(mockContext.executeActionCalls[0].payload.url).to.equal('/external');
+                done();
+            }, 10);
+        });
+
+        it('context.executeAction not called for href when validate is true but href is not registered', function (done) {
+            var link = ReactTestUtils.renderIntoDocument(
+                <MockAppComponent context={mockContext}>
+                    <NavLink validate={true} href='notregister'/>
+                </MockAppComponent>
+            );
+            ReactTestUtils.Simulate.click(ReactDOM.findDOMNode(link), {button: 0});
+            window.setTimeout(function () {
+                expect(testResult.dispatch).to.equal(undefined);
+                done();
+            }, 10);
+        });
+
+        it('context.executeAction called for routeNames', function (done) {
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
                     <NavLink routeName='foo' />
@@ -239,7 +365,8 @@ describe('NavLink', function () {
                 done();
             }, 10);
         });
-        it ('context.executeAction called for absolute urls from same origin', function (done) {
+
+        it('context.executeAction called for absolute urls from same origin', function (done) {
             var navParams = {a: 1, b: true};
             var origin = window.location.origin;
             var link = ReactTestUtils.renderIntoDocument(
@@ -257,6 +384,7 @@ describe('NavLink', function () {
                 done();
             }, 10);
         });
+
         it('context.executeAction not called for external urls', function (done) {
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
@@ -269,6 +397,20 @@ describe('NavLink', function () {
                 done();
             }, 10);
         });
+
+        it('context.executeAction not called for external urls when validate is true', function (done) {
+            var link = ReactTestUtils.renderIntoDocument(
+                <MockAppComponent context={mockContext}>
+                    <NavLink validate={true} href='http://domain.does.not.exist/foo' />
+                </MockAppComponent>
+            );
+            ReactTestUtils.Simulate.click(ReactDOM.findDOMNode(link), {button: 0});
+            window.setTimeout(function () {
+                expect(testResult.dispatch).to.equal(undefined);
+                done();
+            }, 10);
+        });
+
         it('context.executeAction not called for # urls', function (done) {
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
@@ -295,6 +437,7 @@ describe('NavLink', function () {
                 done();
             }, 10);
         });
+
         it('context.executeAction called if followLink=false', function (done) {
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
@@ -319,7 +462,7 @@ describe('NavLink', function () {
                 };
             });
 
-            it ('should not call context.executeAction when a user does not confirm the onbeforeunload method', function (done) {
+            it('should not call context.executeAction when a user does not confirm the onbeforeunload method', function (done) {
                 var link = ReactTestUtils.renderIntoDocument(
                     <MockAppComponent context={mockContext}>
                         <NavLink href='/foo' followLink={false} />
@@ -416,7 +559,7 @@ describe('NavLink', function () {
         it('should update active state', function (done) {
             var link = ReactTestUtils.renderIntoDocument(
                 <MockAppComponent context={mockContext}>
-                    <NavLink href='/foo'>
+                    <NavLink href='/foo' activeClass='active'>
                         bar
                     </NavLink>
                 </MockAppComponent>
@@ -438,12 +581,27 @@ describe('NavLink', function () {
         });
     });
 
+    describe('componentDidMount', function () {
+        it('should only listen if there is an active property', function () {
+            var div = document.createElement('div');
+            ReactDOM.render(
+                <MockAppComponent context={mockContext}>
+                    <NavLink href='/foo' />
+                </MockAppComponent>
+                , div);
+            var routeStore = mockContext.getStore('RouteStore');
+            expect(routeStore.listeners('change').length).to.equal(1);
+            ReactDOM.unmountComponentAtNode(div);
+            expect(routeStore.listeners('change').length).to.equal(0);
+        });
+    });
+
     describe('componentWillUnmount', function () {
         it('should remove the change listener', function () {
             var div = document.createElement('div');
             ReactDOM.render(
                 <MockAppComponent context={mockContext}>
-                    <NavLink href='/foo' />
+                    <NavLink href='/foo' activeClass='active' />
                 </MockAppComponent>
             , div);
             var routeStore = mockContext.getStore('RouteStore');